requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-anymail==8.2
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-countries==7.0
django-rest-swagger==2.2.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-stubs==1.3.0
idna==2.10
importlib-metadata==3.4.0
itypes==1.2.0
Jinja2==2.11.2
kaleido==0.1.0
kombu==5.0.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
pkg-resources==0.0.0
plotly==4.14.3
prompt-toolkit==3.0.14
psutil==5.8.0
psycopg2-binary==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
redis==3.5.3
requests==2.25.1
retrying==1.3.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
